+++ NEWS.md
@@ -13,6 +13,7 @@ This may have been addressed at 64a308ba50b9119108e8946737460de5997b805b by addi
 
 # vcfR 1.10.0
 Released on CRAN 202X-XX-XX
+* Handled deprecated "dplyr::verb_" function in vcfR2tidy
 
 # vcfR 1.9.0
 Released on CRAN 2020-01-10